pyproject.toml
[project]
name = "airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow>=3.0.6",
    "apache-airflow-providers-mongo>=5.2.2",
    "dotenv>=0.9.9",
    "exponent-server-sdk>=2.2.0",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.42.1",
    "google-cloud-bigquery>=3.0.0",
    "numpy>=2.3.2",
    "rollbar>=1.3.0",
]
